# mrcomic-api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mr.Comic API",
    "description": "API для сторонних разработчиков платформы Mr.Comic",
    "version": "1.0.0",
    "contact": {
      "name": "Mr.Comic API Support",
      "email": "api-support@mrcomic.com",
      "url": "https://developer.mrcomic.com"
    }
  },
  "servers": [
    {
      "url": "https://api.mrcomic.com/v1",
      "description": "Продакшн сервер"
    },
    {
      "url": "https://api-sandbox.mrcomic.com/v1",
      "description": "Песочница для тестирования"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/comics": {
      "get": {
        "summary": "Получение списка комиксов",
        "description": "Возвращает список комиксов с пагинацией и возможностью фильтрации",
        "operationId": "getComics",
        "tags": ["Comics"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Количество элементов на странице",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "genre",
            "in": "query",
            "description": "Фильтр по жанру",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Поисковый запрос",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Поле для сортировки",
            "schema": {
              "type": "string",
              "enum": ["title", "releaseDate", "popularity", "rating"],
              "default": "popularity"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Порядок сортировки",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComicsList"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/comics/{id}": {
      "get": {
        "summary": "Получение информации о комиксе",
        "description": "Возвращает детальную информацию о комиксе по его ID",
        "operationId": "getComicById",
        "tags": ["Comics"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID комикса",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comic"
                }
              }
            }
          },
          "404": {
            "description": "Комикс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/comics/{id}/chapters": {
      "get": {
        "summary": "Получение списка глав комикса",
        "description": "Возвращает список глав для указанного комикса",
        "operationId": "getComicChapters",
        "tags": ["Comics"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID комикса",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Количество элементов на странице",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChaptersList"
                }
              }
            }
          },
          "404": {
            "description": "Комикс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/analytics/reading-stats": {
      "get": {
        "summary": "Получение статистики чтения",
        "description": "Возвращает статистику чтения для текущего пользователя",
        "operationId": "getReadingStats",
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "Период статистики",
            "schema": {
              "type": "string",
              "enum": ["day", "week", "month", "year", "all"],
              "default": "month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingStats"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/recommendations": {
      "get": {
        "summary": "Получение рекомендаций",
        "description": "Возвращает персонализированные рекомендации комиксов для текущего пользователя",
        "operationId": "getRecommendations",
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Количество рекомендаций",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsList"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Comic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "format": "uri"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "releaseDate": {
            "type": "string",
            "format": "date"
          },
          "rating": {
            "type": "number",
            "format": "float"
          },
          "chaptersCount": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": ["ongoing", "completed", "hiatus"]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publisher": {
            "type": "string"
          }
        },
        "required": ["id", "title", "coverUrl", "genres", "status"]
      },
      "ComicsList": {
        "type": "object",
        "properties": {
          "comics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comic"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": ["comics", "pagination"]
      },
      "Chapter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "comicId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "number": {
            "type": "number"
          },
          "releaseDate": {
            "type": "string",
            "format": "date"
          },
          "pagesCount": {
            "type": "integer"
          },
          "thumbnailUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": ["id", "comicId", "number", "pagesCount"]
      },
      "ChaptersList": {
        "type": "object",
        "properties": {
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Chapter"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": ["chapters", "pagination"]
      },
...
      "ReadingStats": {
        "type": "object",
        "properties": {
          "totalReadingTime": {
            "type": "integer",
            "description": "Общее время чтения в минутах"
          },
          "comicsRead": {
            "type": "integer",
            "description": "Количество прочитанных комиксов"
          },
          "chaptersRead": {
            "type": "integer",
            "description": "Количество прочитанных глав"
          },
          "pagesRead": {
            "type": "integer",
            "description": "Количество прочитанных страниц"
          },
          "averageReadingSpeed": {
            "type": "number",
            "format": "float",
            "description": "Средняя скорость чтения (страниц в минуту)"
          },
          "mostReadGenres": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "genre": {
                  "type": "string"
                },
                "count": {
                  "type": "integer"
                }
              }
            }
          },
          "readingByDay": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "format": "date"
                },
                "minutes": {
                  "type": "integer"
                },
                "pages": {
                  "type": "integer"
                }
              }
            }
          }
        },
        "required": ["totalReadingTime", "comicsRead", "chaptersRead", "pagesRead"]
      },
      "Recommendation": {
        "type": "object",
        "properties": {
          "comic": {
            "$ref": "#/components/schemas/Comic"
          },
          "score": {
            "type": "number",
            "format": "float",
            "description": "Оценка релевантности рекомендации (0-1)"
          },
          "reason": {
            "type": "string",
            "description": "Причина рекомендации"
          }
        },
        "required": ["comic", "score"]
      },
      "RecommendationsList": {
        "type": "object",
        "properties": {
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            }
          }
        },
        "required": ["recommendations"]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          }
        },
        "required": ["total", "page", "limit", "pages"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        },
        "required": ["code", "message"]
      }
    }
  }
}